==> code/analysis/1_exploratory_timeseries.qmd <==
---
title: "Exploratory - Time series"
author: "Leah Lariscy"
format: html
editor: visual
toc: true
toc-depth: 4
df-print: paged
code-overflow: wrap
---

## Load packages

```{r}
knitr::opts_chunk$set(message=F)
```

```{r,message=FALSE}
library(tidyverse)
library(here)
library(fpp3)
library(ggthemes)
```

## Load data

```{r}
data <- readRDS(here("data/processed_data/covid_wbe_7dma_weekly.rds"))
#from pre_processing script

data
```

## Look at time series

### 1. Viral Load

```{r}
data %>% na.omit() %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=log10(vl_7dma),color="VL 7DMA")) +
  geom_line(aes(y=log10(avg_vl),color="VL")) +
  theme_bw() +
  labs(title = "Viral Load vs Viral Load 7DMA") +
  ylab("Log10 Weekly Average Copies per Day")

```

### 2. qPCR positivity rate

```{r}
data %>% na.omit() %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=avg_pos_rate,color="WW PR")) +
  geom_line(aes(y=pr_7dma,color="WW PR 7DMA")) +
  theme_bw() +
  labs(title = "Wastewater Pos Rate vs Wastewater Pos Rate 7DMA") +
  ylab("Weekly Average Percent Positive qPCR Reactions")
```

### 3. Case symptom onset

```{r}
data %>% na.omit() %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases.symptom.onset,color="Symptom onset")) +
  geom_line(aes(y=symptom_onset_7dma,color="Symptom onset 7DMA")) +
  theme_bw() +
  labs(title = "Symptom Onset vs Symptom Onset 7DMA") +
  ylab("Weekly Average Case Symptom Onset")
```

### 4. Cases reported

```{r}
data %>% na.omit() %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases.reported,color="Cases reported")) +
  geom_line(aes(y=case_report_7dma,color="Cases reported 7DMA")) +
  theme_bw() +
  labs(title = "Cases Reported vs Cases Reported 7DMA") +
  ylab("Weekly Average Cases Reported")
```

### 5. Case test positives

```{r}
data %>% na.omit() %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=pcr_pos,color="Pos clinical tests")) +
  geom_line(aes(y=case_pcr_pos_7dma,color="Pos clinical tests 7DMA")) +
  geom_line(aes(y=pcr_pos_7dma,color="7DMA calculated by DPH")) +
  theme_bw() +
  labs(title = "Positive Clinical Tests vs Positive Clinical Test 7DMA") +
  ylab("Weekly Average Positive PCR Tests")
```

### 6. Case test positivity rate

```{r}
data %>% na.omit() %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=prop_pos*100,color="Clinical Test PR")) +
  geom_line(aes(y=case_prop_pos_7dma,color="Clinical Test PR 7DMA")) +
  theme_bw() +
  labs(title = "Clinical Positivity Rate vs Clinical Positivity Rate 7DMA") +
  ylab("Weekly Average Percent Positive PCR Tests")

```

### 9. Clinical test admin

```{r}
data %>% na.omit() %>% 
  ggplot(aes(week, pcr_tests)) +
  geom_line() +
  theme_bw() +
  labs(title = "Clinical Tests Administered Over Time") +
  ylab("Weekly Average Clinical Tests")
```

### Viral Load, Detection Freq, Case Pos Rate

```{r}
data %>% na.omit() %>% 
  ggplot(aes(week)) +
  geom_area(aes(y=avg_pos_rate),fill="#2A42DB",alpha=.5) +
  geom_area(aes(y=prop_pos*100),alpha=.8,fill="#610EAB") +
  theme_clean() +
  labs(title = "Wastewater Detection Frequency and Test Positivity Rate Over Time") +
  ylab("Percent Positive") +
  theme(axis.text.x = element_text(size=10),
        axis.title.x.bottom = element_text(size=15,face = "bold"),
        axis.text.y = element_text(size=10),
        axis.title.y.left = element_text(size=15,face = "bold")) +
    scale_fill_manual("Time Series", values=c("#2A42DB", "#610EAB"),
                        breaks=c("Wastewater","Clinical"))
```

```{r}
data %>% mutate(prop_pos=prop_pos*100) %>% 
  pivot_longer(cols = c(avg_pos_rate,prop_pos),
                      names_to = "pos_rate") %>% 
  ggplot(aes(week, value, fill=pos_rate)) +
  geom_area() +
  theme_clean() +
  labs(title = "Wastewater Detection Frequency and Test Positivity Rate Over Time") +
  ylab("Percent Positive") +
  xlab("Week") +
  theme(axis.text.x = element_text(size=10),
        axis.title.x.bottom = element_text(size=15,face = "bold"),
        axis.text.y = element_text(size=10),
        axis.title.y.left = element_text(size=15,face = "bold")) +
    scale_fill_brewer(name="Time Series", palette="Set1",
                        labels=c("Wastewater","Clinical"))

ggsave(here("figures/wastewater_clinical_timeseries.png"))
```

## Notes

-   all time series follow roughly the same trend

-   peaks happen roughly twice per year, the first during the fall (Aug-Oct, potentially influenced by student population influx) and second during the winter (Dec-Feb)

-   tests reported, in general, peak during peak transmission times, but that trend does not continue past the beginning of 2022 (see 9. Test admin), indicating the shift to at-home testing and subsequent under-reporting

-   it may make sense to split at March-April 2022 as correlations between wastewater metrics and case metrics could weaken past this point

```{r}
ggpubr::ggscatter(data, x="avg_pos_rate", y="cases.reported",
          add="reg.line",conf.int = TRUE,
          cor.coef = TRUE, cor.method = "pearson")
  
```

```{r}
ggpubr::ggscatter(data, x="avg_pos_rate", y="prop_pos",
          add="reg.line",conf.int = TRUE,
          cor.coef = TRUE, cor.method = "pearson")

ggsave(here("figures/DetectionFreq_CasePosRate.png"))
```
